rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.0
google-auth==2.25.2
google-cloud-storage==2.14.0
chromadb==0.5.0
numpy==1.26.0
requests==2.31.0
httpx==0.25.2
PyPDF2==2.12.1
